// src/models/Enums.ts
export enum Actions {
    ERROR = "ERROR",
    INFO = "INFO",
    USERNAME_CHANGED = "USERNAMECHANGED",
    PASSWORD_CHANGED = "PASSWORDCHANGED",
    LOGGED_IN = "LOGGEDIN",
    LOGGED_OUT = "LOGGEDOUT",
    POST_CLEARED = "POSTCLEARED",
    TITLE_CHANGED = "TITLECHANGED",
    BODY_CHANGED = "BODYCHANGED",
    POST_CREATED = "POSTCREATED",
    POST_UPDATED = "POSTUPDATED",
    POSTS_FETCHED = "POSTSFETCHED",
    POST_EDITED = "POSTEDITED",
    POST_DELETED = "POSTDELETED"
};